refactor: modify setRecords logic
