+++ packages/vtable/src/PivotTable.ts
@@ -66,12 +66,19 @@ export class PivotTable extends BaseTable implements PivotTableAPI {
       options.indicatorsAsCol && !options.columns?.length && !options.columnTree ? [] : cloneDeep(options.columnTree);
     this.internalProps.rowTree =
       !options.indicatorsAsCol && !options.rows?.length && !options.rowTree ? [] : cloneDeep(options.rowTree);
+    this.internalProps.records = options.records;
+    this.pivotSortState = [];
+    if (options.pivotSortState) {
+      this.updatePivotSortState(options.pivotSortState);
+    }
     //分页配置
     this.pagination = options.pagination;
     this.internalProps.columnResizeType = options.columnResizeType ?? 'column';
     this.internalProps.dataConfig = cloneDeep(options.dataConfig);
+
     this.internalProps.enableDataAnalysis = options.enableDataAnalysis;
-    if (this.internalProps.enableDataAnalysis) {
+    const records = this.internalProps.records;
+    if (this.internalProps.enableDataAnalysis && (options.rows || options.columns)) {
       const rowKeys =
         options.rows?.reduce((keys, rowObj) => {
           if (typeof rowObj === 'string') {
@@ -114,23 +121,34 @@ export class PivotTable extends BaseTable implements PivotTableAPI {
         this.internalProps.columnTree, //传递自定义树形结构会在dataset中补充指标节点children
         this.internalProps.rowTree
       );
-    }
-
-    this.refreshHeader();
-
-    this.pivotSortState = [];
-    if (options.pivotSortState) {
-      this.updatePivotSortState(options.pivotSortState);
-    }
-
-    // if (options.dataSource) {
-    //   _setDataSource(this, options.dataSource);
-    // } else
-    if (options.records) {
-      this.setRecords(options.records as any, this.internalProps.sortState);
+      this.internalProps.layoutMap = new PivotHeaderLayoutMap(this, this.dataset);
+    } else if (Array.isArray(this.internalProps.columnTree) || Array.isArray(this.internalProps.rowTree)) {
+      this.internalProps.layoutMap = new PivotHeaderLayoutMap(this, null);
+      //判断如果数据是二维数组 则标识已经分析过 直接从二维数组挨个读取渲染即可
+      //不是二维数组 对应是个object json对象 则表示flat数据，需要对应行列维度进行转成方便数据查询的行列树结构
+      if (records?.[0]?.constructor !== Array) {
+        this.flatDataToObjects = new FlatDataToObjects(
+          {
+            rows: this.internalProps.layoutMap.fullRowDimensionKeys,
+            columns: this.internalProps.layoutMap.colDimensionKeys,
+            indicators: this.internalProps.layoutMap.indicatorKeys,
+            indicatorsAsCol: this.internalProps.layoutMap.indicatorsAsCol,
+            indicatorDimensionKey: this.internalProps.layoutMap.indicatorDimensionKey
+          },
+          records
+        );
+      }
     } else {
-      this.setRecords([]);
+      console.warn('your option is invalid, please check it!');
+      return;
     }
+    this.refreshHeader();
+    this.stateManeger.initCheckedState(records);
+    // this.internalProps.frozenColCount = this.options.frozenColCount || this.rowHeaderLevelCount;
+    // 生成单元格场景树
+    this.scenegraph.createSceneGraph();
+    // this.render();
+
     if (options.title) {
       this.internalProps.title = new Title(options.title, this);
       this.scenegraph.resize();
@@ -173,6 +191,12 @@ export class PivotTable extends BaseTable implements PivotTableAPI {
       options.indicatorsAsCol && !options.columns?.length && !options.columnTree ? [] : cloneDeep(options.columnTree);
     this.internalProps.rowTree =
       !options.indicatorsAsCol && !options.rows?.length && !options.rowTree ? [] : cloneDeep(options.rowTree);
+    options.records && (this.internalProps.records = options.records);
+    this.stateManeger.initCheckedState(this.internalProps.records);
+    this.pivotSortState = [];
+    if (options.pivotSortState) {
+      this.updatePivotSortState(options.pivotSortState);
+    }
     //分页配置
     this.pagination = options.pagination;
     // 更新protectedSpace
@@ -196,10 +220,10 @@ export class PivotTable extends BaseTable implements PivotTableAPI {
         }
       });
     }
-
+    const records = this.internalProps.records;
     //TODO 这里需要加上判断 dataConfig是否有配置变化
     if (this.internalProps.enableDataAnalysis && (options.rows || options.columns)) {
-      const rowKeys = options.rows.reduce((keys, rowObj) => {
+      const rowKeys = options.rows?.reduce((keys, rowObj) => {
         if (typeof rowObj === 'string') {
           keys.push(rowObj);
         } else {
@@ -207,7 +231,7 @@ export class PivotTable extends BaseTable implements PivotTableAPI {
         }
         return keys;
       }, []);
-      const columnKeys = options.columns.reduce((keys, columnObj) => {
+      const columnKeys = options.columns?.reduce((keys, columnObj) => {
         if (typeof columnObj === 'string') {
           keys.push(columnObj);
         } else {
@@ -236,7 +260,28 @@ export class PivotTable extends BaseTable implements PivotTableAPI {
         this.internalProps.columnTree, //传递自定义树形结构会在dataset中补充指标节点children
         this.internalProps.rowTree
       );
+      internalProps.layoutMap = new PivotHeaderLayoutMap(this, this.dataset);
+    } else if (Array.isArray(this.internalProps.columnTree) || Array.isArray(this.internalProps.rowTree)) {
+      internalProps.layoutMap = new PivotHeaderLayoutMap(this, null);
+      //判断如果数据是二维数组 则标识已经分析过 直接从二维数组挨个读取渲染即可
+      //不是二维数组 对应是个object json对象 则表示flat数据，需要对应行列维度进行转成方便数据查询的行列树结构
+      if (records?.[0]?.constructor !== Array) {
+        this.flatDataToObjects = new FlatDataToObjects(
+          {
+            rows: internalProps.layoutMap.fullRowDimensionKeys,
+            columns: internalProps.layoutMap.colDimensionKeys,
+            indicators: internalProps.layoutMap.indicatorKeys,
+            indicatorsAsCol: internalProps.layoutMap.indicatorsAsCol,
+            indicatorDimensionKey: internalProps.layoutMap.indicatorDimensionKey
+          },
+          records
+        );
+      }
+    } else {
+      console.warn('your option is invalid, please check it!');
+      return;
     }
+
     // 更新表头
     this.refreshHeader();
 
@@ -249,26 +294,24 @@ export class PivotTable extends BaseTable implements PivotTableAPI {
     // // 恢复selection状态
     // internalProps.selection.range = range;
     // this._updateSize();
-    // 传入新数据
-    // if (options.dataSource) {
-    //   _setDataSource(this, options.dataSource);
-    // } else
-    if (options.records) {
-      this.setRecords(options.records as any, undefined);
-    } else {
-      this._resetFrozenColCount();
-      // 生成单元格场景树
-      this.scenegraph.createSceneGraph();
-      this.render();
-    }
+
+    // 清空单元格内容
+    this.scenegraph.clearCells();
+    // this.internalProps.frozenColCount = this.options.frozenColCount || this.rowHeaderLevelCount;
+    // 生成单元格场景树
+    this.scenegraph.createSceneGraph();
+
+    // if (this.internalProps.title && !this.internalProps.title.isReleased) {
+    //   this._updateSize();
+    //   this.internalProps.title.resize();
+    //   this.scenegraph.resize();
+    // }
     if (options.title) {
       this.internalProps.title = new Title(options.title, this);
       this.scenegraph.resize();
     }
-    this.pivotSortState = [];
-    if (options.pivotSortState) {
-      this.updatePivotSortState(options.pivotSortState);
-    }
+
+    // this.render();
     return new Promise(resolve => {
       setTimeout(resolve, 0);
     });
@@ -320,26 +363,6 @@ export class PivotTable extends BaseTable implements PivotTableAPI {
     //原表头绑定的事件 解除掉
     if (internalProps.headerEvents) {
       internalProps.headerEvents.forEach((id: number) => this.off(id));
-    }
-    const records = this.options.records ?? this.internalProps.records;
-    if (this.options.enableDataAnalysis) {
-      internalProps.layoutMap = new PivotHeaderLayoutMap(this, this.dataset);
-    } else if (Array.isArray(this.internalProps.columnTree) || Array.isArray(this.internalProps.rowTree)) {
-      internalProps.layoutMap = new PivotHeaderLayoutMap(this, null);
-      //判断如果数据是二维数组 则标识已经分析过 直接从二维数组挨个读取渲染即可
-      //不是二维数组 对应是个object json对象 则表示flat数据，需要对应行列维度进行转成方便数据查询的行列树结构
-      if (records?.[0]?.constructor !== Array) {
-        this.flatDataToObjects = new FlatDataToObjects(
-          {
-            rows: internalProps.layoutMap.fullRowDimensionKeys,
-            columns: internalProps.layoutMap.colDimensionKeys,
-            indicators: internalProps.layoutMap.indicatorKeys,
-            indicatorsAsCol: internalProps.layoutMap.indicatorsAsCol,
-            indicatorDimensionKey: internalProps.layoutMap.indicatorDimensionKey
-          },
-          records
-        );
-      }
     }
 
     //设置列宽
@@ -485,8 +508,18 @@ export class PivotTable extends BaseTable implements PivotTableAPI {
         ? fieldFormat(valueNode?.record, col, row, this)
         : valueNode?.value ?? '';
     }
-    const { field, fieldFormat } = this.internalProps.layoutMap.getBody(col, row);
-    return this.getFieldData(fieldFormat || field, col, row);
+    const { fieldFormat } = this.internalProps.layoutMap.getBody(col, row);
+    const rowIndex = this.getBodyIndexByRow(row);
+    const colIndex = this.getBodyIndexByCol(col);
+    const dataValue = this.records[rowIndex]?.[colIndex];
+    const cellHeaderPaths = this.internalProps.layoutMap.getCellHeaderPaths(col, row);
+
+    if (typeof fieldFormat === 'function') {
+      const fieldResult = fieldFormat({ dataValue, ...cellHeaderPaths }, col, row, this);
+      return fieldResult;
+    }
+    return dataValue;
+    // return this.getFieldData(fieldFormat || field, col, row);
   }
 
   getCellOriginValue(col: number, row: number): FieldData {
@@ -526,8 +559,12 @@ export class PivotTable extends BaseTable implements PivotTableAPI {
       );
       return treeNode?.value;
     }
-    const { field } = table.internalProps.layoutMap.getBody(col, row);
-    return table.getFieldData(field, col, row);
+    const rowIndex = this.getBodyIndexByRow(row);
+    const colIndex = this.getBodyIndexByCol(col);
+    const dataValue = this.records[rowIndex]?.[colIndex];
+    return dataValue;
+    // const { field } = table.internalProps.layoutMap.getBody(col, row);
+    // return table.getFieldData(field, col, row);
   }
 
   // 获取原始数据
@@ -567,7 +604,10 @@ export class PivotTable extends BaseTable implements PivotTableAPI {
       );
       return treeNode?.record;
     }
-    return undefined;
+    const rowIndex = this.getBodyIndexByRow(row);
+    const colIndex = this.getBodyIndexByCol(col);
+    const dataValue = this.records[rowIndex]?.[colIndex];
+    return dataValue;
   }
   /**
    * 全量更新排序规则 TODO  待完善
@@ -795,5 +835,34 @@ export class PivotTable extends BaseTable implements PivotTableAPI {
       isPivotCorner: this.isCornerHeader(col, row)
     };
     return result;
+  }
+
+  /**
+   * 设置表格数据 及排序状态
+   * @param records
+   * @param sort
+   */
+  setRecords(records: Array<any>): void {
+    // const time = typeof window !== 'undefined' ? window.performance.now() : 0;
+    // // 清空单元格内容
+    // this.scenegraph.clearCells();
+    // this.internalProps.records = records;
+    // if (records) {
+    //   // _setRecords(this, records);
+    //   this.refreshRowColCount();
+    // } else {
+    //   // _setRecords(this, records);
+    // }
+    // this.stateManeger.initCheckedState(records);
+    // this.internalProps.frozenColCount = this.options.frozenColCount || this.rowHeaderLevelCount;
+    // // 生成单元格场景树
+    // this.scenegraph.createSceneGraph();
+    // if (this.internalProps.title && !this.internalProps.title.isReleased) {
+    //   this._updateSize();
+    //   this.internalProps.title.resize();
+    //   this.scenegraph.resize();
+    // }
+    // this.render();
+    // console.log('setRecords cost time:', (typeof window !== 'undefined' ? window.performance.now() : 0) - time);
   }
 }
